Squashed commit of the following:
    commit 4f66f7d6d44daaa1ec2e1b87991961295c3ca4e3
    Author: coanor <coanor@gmail.com>
    Date:   Mon Jan 27 15:56:50 2014 +0800

        add tns module

    commit 471a9d48bc8d27297c88d2810e5b48d09f90dcb6
    Author: coanor <coanor@gmail.com>
    Date:   Mon Jan 27 10:04:18 2014 +0800

    Squashed commit of the following:

            commit 8b7e2a3f77420c0281f1926c25536f4d4e2667e9
            Author: coanor <coanor@gmail.com>
            Date:   Thu Jan 23 17:31:14 2014 +0800

                add tds protocol

            commit f21f288d41738ee2ecca1c9703870f38722412a7
            Author: coanor <coanor@gmail.com>
            Date:   Tue Jan 21 17:57:30 2014 +0800

                add log module for tds
                debuging app-layer-tds*

            commit be13f674e67b8c59b812912d4f165751dc82afa4
            Author: coanor <coanor@gmail.com>
            Date:   Mon Jan 20 17:47:51 2014 +0800

                add tds module

            commit afa5a30296623276f9dc4a6db17aa6050ff7a78f
            Author: coanor <coanor@gmail.com>
            Date:   Wed Jan 15 16:35:58 2014 +0800

                updte:
                    add following keywords support
                         * date         (format is `year-month-day')
                         * weekdays     (only support 1-7 currently)
                         * time         (format is `hour:minute:second', and is 24-hours format)
                         * monthdays    (format is 1~31)

+++ src/app-layer-mysql-common.h
@@ -5,8 +5,6 @@
 #include "app-layer-protos.h"
 #include "app-layer-parser.h"
 #include "app-layer-detect-proto.h"
-#include "app-layer-mysql.h"
-#include "app-layer-mysql-common.h"
 #include "util-byte.h"
 #include "queue.h"
 
@@ -117,7 +115,7 @@ typedef struct MysqlPktHeader_ {
 typedef struct MysqlClientCommand_ {
     MysqlPktHeader hdr;
     MysqlCommand cmd;
-    int sql_size;
+    uint16_t sql_size;
     char *sql;
 } MysqlClientCommand;
 
@@ -154,7 +152,6 @@ typedef struct PendingPkt_ {
  * \brief MySQL transaction, request/reply with same TX id
  */
 typedef struct MysqlTransaction_ {
-    /* both not in use */
     uint16_t tx_num;
     uint16_t tx_id;
